requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
PyJWT==2.9.0
bcrypt==4.2.0
openai==1.51.0
alembic==1.13.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-multipart==0.0.12
python-dotenv==1.0.1
